backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
attrs==25.1.0
azure-ai-translation-text==1.0.1
azure-cognitiveservices-speech==1.42.0
azure-core==1.32.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
cartesia==1.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
dacite==1.9.2
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
duckdb==1.2.0
et_xmlfile==2.0.0
Faker==19.13.0
fastapi==0.115.9
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gdown==5.2.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gotrue==2.12.2
gradio==5.19.0
gradio_client==1.7.2
greenlet==3.1.1
groq==0.18.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
htmlmin==0.1.12
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.1
iniconfig==2.1.0
isodate==0.7.2
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kaleido==0.2.1
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.58
langchain-experimental==0.3.4
langchain-google-genai==2.1.3
langchain-groq==0.2.4
langchain-openai==0.3.16
langchain-openrouter==0.0.1
langchain-text-splitters==0.3.6
langsmith==0.3.11
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
mypy-extensions==1.0.0
narwhals==1.38.1
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas>=2.0.3
pandasai
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
plotly==6.0.1
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pydyf==0.11.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyphen==0.17.2
PySocks==1.7.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
RapidFuzz==3.13.0
realtime==2.5.3
regex==2024.11.6
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
ruff==0.9.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==4.1.0
setuptools==75.8.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.38
sqlglot==25.34.1
sqlglotrs==0.3.0
starlette==0.45.3
statsmodels==0.14.4
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.4.1
tqdm==4.67.1
transformers==4.52.4
typeguard==4.4.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
visions==0.7.6
watchfiles==1.0.5
weasyprint==65.1
webencodings==0.5.1
websockets==15.0
Werkzeug==3.1.3
wordcloud==1.9.4
yarl==1.18.3
ydata-profiling==4.12.2
zopfli==0.2.3.post1
zstandard==0.23.0

# Predictive Analytics Libraries
prophet==1.1.5
xgboost==2.0.3
lightgbm==4.3.0
pmdarima==2.0.4
dateparser>=1.1.0

# Knowledge Base Document Processing
pdfplumber==0.11.4          # PDF text and table extraction
python-docx==1.1.2          # DOCX reading
pgvector==0.3.6             # PostgreSQL vector extension client
